-refactor package name -some fixes
