+++ ariss-log-action/src/main/java/ru/unit/techno/ariss/log/action/lib/api/LogAction.java
@@ -1,7 +1,7 @@
 
-package ru.unit.techno.arris.log.action.lib.api;
+package ru.unit.techno.ariss.log.action.lib.api;
 
-import ru.unit.techno.arris.log.action.lib.model.ActionObject;
+import ru.unit.techno.ariss.log.action.lib.model.ActionObject;
 
 public interface LogAction {
     void logSuccessAction(ActionObject actionObject);
