+++ ariss-log-action/src/main/java/ru/unit/techno/ariss/log/action/lib/config/DeviceEventConfig.java
@@ -1,11 +1,11 @@
 
-package ru.unit.techno.arris.log.action.lib.config;
+package ru.unit.techno.ariss.log.action.lib.config;
 
 import lombok.Data;
 import lombok.Getter;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.stereotype.Component;
-import ru.unit.techno.arris.log.action.lib.model.MetaObject;
+import ru.unit.techno.ariss.log.action.lib.model.MetaObject;
 
 import java.util.Map;
