+++ ariss-log-action/src/main/java/ru/unit/techno/ariss/log/action/lib/config/LogActionConfig.java
@@ -1,10 +1,10 @@
 
-package ru.unit.techno.arris.log.action.lib.config;
+package ru.unit.techno.ariss.log.action.lib.config;
 
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-@ComponentScan(value = "ru.unit.techno.arris.log.action.lib")
+@ComponentScan(value = "ru.unit.techno.ariss.log.action.lib")
 public class LogActionConfig {
 }
