+++ ariss-log-action/src/main/java/ru/unit/techno/ariss/log/action/lib/entity/Description.java
@@ -1,4 +1,4 @@
-package ru.unit.techno.arris.log.action.lib.entity;
+package ru.unit.techno.ariss.log.action.lib.entity;
 
 import lombok.Data;
 import lombok.experimental.Accessors;
