+++ ariss-log-action/src/main/java/ru/unit/techno/ariss/log/action/lib/repository/EventRepository.java
@@ -1,8 +1,8 @@
-package ru.unit.techno.arris.log.action.lib.repository;
+package ru.unit.techno.ariss.log.action.lib.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import ru.unit.techno.arris.log.action.lib.entity.Event;
+import ru.unit.techno.ariss.log.action.lib.entity.Event;
 
 @Repository
 public interface EventRepository extends JpaRepository<Event, Long> {
