+++ ariss-log-action/src/main/java/ru/unit/techno/ariss/log/action/lib/service/LogActionBuilderImpl.java
@@ -1,13 +1,13 @@
 
-package ru.unit.techno.arris.log.action.lib.service;
+package ru.unit.techno.ariss.log.action.lib.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
-import ru.unit.techno.arris.log.action.lib.api.LogAction;
-import ru.unit.techno.arris.log.action.lib.api.LogActionBuilder;
-import ru.unit.techno.arris.log.action.lib.entity.Description;
-import ru.unit.techno.arris.log.action.lib.model.ActionObject;
-import ru.unit.techno.arris.log.action.lib.model.ActionStatus;
+import ru.unit.techno.ariss.log.action.lib.api.LogAction;
+import ru.unit.techno.ariss.log.action.lib.api.LogActionBuilder;
+import ru.unit.techno.ariss.log.action.lib.model.ActionObject;
+import ru.unit.techno.ariss.log.action.lib.model.ActionStatus;
+import ru.unit.techno.ariss.log.action.lib.entity.Description;
 
 import java.time.LocalDateTime;
 
@@ -24,7 +24,7 @@ public class LogActionBuilderImpl implements LogActionBuilder {
                                               ActionStatus actionStatus) {
         logAction.logSuccessAction(new ActionObject()
                 .setDeviceId(deviceId)
-                .setDeviceId(deviceId)
+                .setCommonId(commonId)
                 .setActionStatus(actionStatus)
                 .setEventTime(LocalDateTime.now())
                 .setGosNumber(gosNumber));
