+++ ariss-log-action/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,2 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-  ru.unit.techno.arris.log.action.lib.config.LogActionConfig
+  ru.unit.techno.ariss.log.action.lib.config.LogActionConfig
